% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessioninfo2renv-package.R
\docType{package}
\name{sessioninfo2renv-package}
\alias{sessioninfo2renv}
\alias{sessioninfo2renv-package}
\title{sessioninfo2renv: Convert session info to a Lockfile}
\description{
Convert session info (base or tidyverse version) object or printed output into an 'renv' lockfile.
}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})

}
\keyword{internal}
