% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unformat.R
\name{unformat_session_info}
\alias{unformat_session_info}
\title{Get a \code{session_info} object from its printed representation}
\usage{
unformat_session_info(file)
}
\arguments{
\item{file}{A file path to a text file containing the output of}
}
\description{
Get a \code{session_info} object from its printed representation
}
\examples{
file <- system.file("extdata", "sessioninfo.txt", package = "sessioninfo")

}
