% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_lockfile.R
\name{as_lockfile}
\alias{as_lockfile}
\title{Convert an object to an renv lockfile}
\usage{
as_lockfile(x, ...)
}
\arguments{
\item{x}{An object to convert}
}
\description{
Convert an object to an renv lockfile
}
