% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_lockfile.R
\name{as_lockfile.session_info}
\alias{as_lockfile.session_info}
\title{Convert a \code{session_info} object to an renv lockfile}
\usage{
\method{as_lockfile}{session_info}(x, lockfile = stdout(), ...)
}
\arguments{
\item{x}{An object to convert}

\item{lockfile}{A file path to write the lockfile to. If \code{stdout()} (the
default), the lockfile is printed to the console.}

\item{...}{Additional arguments passed to methods}
}
\description{
\code{session_info} objects are produced by the \code{\link[sessioninfo:session_info]{sessioninfo::session_info()}}
function.
}
